requirements.txt
pyTelegramBotAPI==4.13.0
aiohttp==3.8.5
pymongo==4.5.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
CurrencyConverter==0.17.9
Pillow==10.0.0
colorama==0.4.6
emoji==2.8.0
langdetect==1.0.9
translators==5.8.3